SJLSHS_Portal/sjlshs_backend/requirements.txt
anyascii==0.3.1
asgiref==3.5.2
beautifulsoup4==4.11.2
blis==0.7.9
catalogue==2.0.8
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
colorama==0.4.6
confection==0.0.4
crispy-bootstrap5==0.7
cymem==2.0.7
defusedxml==0.7.1
diff-match-patch==20200713
Django==4.1.2
django-adminlte3==0.1.6
django-ajax-selects==2.2.0
django-autocomplete-light==3.9.4
django-crispy-forms==1.14.0
django-filter==22.1
django-formtools==2.4
django-import-export==3.0.2
django-modelcluster==6.0
django-otp==1.1.5
django-permissionedforms==0.1
django-phonenumber-field==6.4.0
django-taggit==3.1.0
django-treebeard==4.6.1
django-two-factor-auth==1.15.1
django-utils-six==2.0
django_debug_toolbar==3.8.1
djangorestframework==3.14.0
draftjs-exporter==2.1.7
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.5.0/en_core_web_sm-3.5.0-py3-none-any.whl
et-xmlfile==1.1.0
html5lib==1.1
idna==3.4
Jinja2==3.1.2
l18n==2021.3
langcodes==3.3.0
MarkupPy==1.14
MarkupSafe==2.1.2
mathparse==0.1.2
murmurhash==1.0.9
mysqlclient==2.1.1
numpy==1.24.1
odfpy==1.4.1
openpyxl==3.0.10
packaging==23.0
pathy==0.10.1
pdf2image==1.16.3
phonenumbers==8.13.7
Pillow==9.3.0
preshed==3.0.8
pydantic==1.10.4
PyMuPDF==1.21.1
pypng==0.20220715.0
python-dateutil==2.8.2
pytz==2022.7.1
PyYAML==6.0
qrcode==7.4.2
requests==2.28.2
six==1.16.0
smart-open==6.3.0
sorl-thumbnail==12.9.0
soupsieve==2.4
spacy==3.5.1
spacy-legacy==3.0.11
spacy-loggers==1.0.4
spacy-lookups-data==1.0.3
SQLAlchemy==1.3.24
sqlparse==0.4.3
srsly==2.4.5
tablib==3.3.0
telepath==0.3
thinc==8.1.9
tqdm==4.64.1
typer==0.7.0
typing_extensions==4.4.0
tzdata==2022.5
urllib3==1.26.14
wagtail==4.2.1
wasabi==1.1.1
webencodings==0.5.1
Willow==1.4.1
xlrd==2.0.1
xlwt==1.3.0
